Also replaced bar chart in YourAnswers
